+++ vmm/src/config.rs
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+pub use crate::vm_config::*;
 use clap::ArgMatches;
-use net_util::MacAddr;
 use option_parser::{
     ByteSized, IntegerList, OptionParser, OptionParserError, StringList, Toggle, Tuple,
 };
@@ -12,7 +12,6 @@ use serde::{Deserialize, Serialize};
 use std::collections::{BTreeSet, HashMap};
 use std::convert::From;
 use std::fmt;
-use std::net::Ipv4Addr;
 use std::path::PathBuf;
 use std::result;
 use std::str::FromStr;
@@ -22,18 +21,8 @@ use virtio_devices::{RateLimiterConfig, TokenBucketConfig};
 pub const DEFAULT_VCPUS: u8 = 1;
 pub const DEFAULT_MEMORY_MB: u64 = 512;
 
-// When booting with PVH boot the maximum physical addressable size
-// is a 46 bit address space even when the host supports with 5-level
-// paging.
-pub const DEFAULT_MAX_PHYS_BITS: u8 = 46;
-
 pub const DEFAULT_RNG_SOURCE: &str = "/dev/urandom";
-pub const DEFAULT_NUM_QUEUES_VUNET: usize = 2;
-pub const DEFAULT_QUEUE_SIZE_VUNET: u16 = 256;
-pub const DEFAULT_NUM_QUEUES_VUBLK: usize = 1;
-pub const DEFAULT_QUEUE_SIZE_VUBLK: u16 = 128;
 
-pub const DEFAULT_NUM_PCI_SEGMENTS: u16 = 1;
 const MAX_NUM_PCI_SEGMENTS: u16 = 16;
 
 /// Errors associated with VM configuration parameters.
@@ -444,12 +433,6 @@ impl<'a> VmParams<'a> {
     }
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub enum HotplugMethod {
-    Acpi,
-    VirtioMem,
-}
-
 impl Default for HotplugMethod {
     fn default() -> Self {
         HotplugMethod::Acpi
@@ -473,29 +456,8 @@ impl FromStr for HotplugMethod {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct CpuAffinity {
-    pub vcpu: u8,
-    pub host_cpus: Vec<u8>,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Eq, Deserialize, Serialize)]
-pub struct CpuFeatures {
-    #[cfg(target_arch = "x86_64")]
-    #[serde(default)]
-    pub amx: bool,
-}
-
 pub enum CpuTopologyParseError {
     InvalidValue(String),
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct CpuTopology {
-    pub threads_per_core: u8,
-    pub cores_per_die: u8,
-    pub dies_per_package: u8,
-    pub packages: u8,
 }
 
 impl FromStr for CpuTopology {
@@ -525,26 +487,6 @@ impl FromStr for CpuTopology {
 
         Ok(t)
     }
-}
-
-fn default_cpuconfig_max_phys_bits() -> u8 {
-    DEFAULT_MAX_PHYS_BITS
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct CpusConfig {
-    pub boot_vcpus: u8,
-    pub max_vcpus: u8,
-    #[serde(default)]
-    pub topology: Option<CpuTopology>,
-    #[serde(default)]
-    pub kvm_hyperv: bool,
-    #[serde(default = "default_cpuconfig_max_phys_bits")]
-    pub max_phys_bits: u8,
-    #[serde(default)]
-    pub affinity: Option<Vec<CpuAffinity>>,
-    #[serde(default)]
-    pub features: CpuFeatures,
 }
 
 impl CpusConfig {
@@ -637,27 +579,6 @@ impl Default for CpusConfig {
     }
 }
 
-fn default_platformconfig_num_pci_segments() -> u16 {
-    DEFAULT_NUM_PCI_SEGMENTS
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct PlatformConfig {
-    #[serde(default = "default_platformconfig_num_pci_segments")]
-    pub num_pci_segments: u16,
-    #[serde(default)]
-    pub iommu_segments: Option<Vec<u16>>,
-    #[serde(default)]
-    pub serial_number: Option<String>,
-    #[serde(default)]
-    pub uuid: Option<String>,
-    #[serde(default)]
-    pub oem_strings: Option<Vec<String>>,
-    #[cfg(feature = "tdx")]
-    #[serde(default)]
-    pub tdx: bool,
-}
-
 impl PlatformConfig {
     pub fn parse(platform: &str) -> Result<Self> {
         let mut parser = OptionParser::new();
@@ -735,51 +656,6 @@ impl Default for PlatformConfig {
             tdx: false,
         }
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct MemoryZoneConfig {
-    pub id: String,
-    pub size: u64,
-    #[serde(default)]
-    pub file: Option<PathBuf>,
-    #[serde(default)]
-    pub shared: bool,
-    #[serde(default)]
-    pub hugepages: bool,
-    #[serde(default)]
-    pub hugepage_size: Option<u64>,
-    #[serde(default)]
-    pub host_numa_node: Option<u32>,
-    #[serde(default)]
-    pub hotplug_size: Option<u64>,
-    #[serde(default)]
-    pub hotplugged_size: Option<u64>,
-    #[serde(default)]
-    pub prefault: bool,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct MemoryConfig {
-    pub size: u64,
-    #[serde(default)]
-    pub mergeable: bool,
-    #[serde(default)]
-    pub hotplug_method: HotplugMethod,
-    #[serde(default)]
-    pub hotplug_size: Option<u64>,
-    #[serde(default)]
-    pub hotplugged_size: Option<u64>,
-    #[serde(default)]
-    pub shared: bool,
-    #[serde(default)]
-    pub hugepages: bool,
-    #[serde(default)]
-    pub hugepage_size: Option<u64>,
-    #[serde(default)]
-    pub prefault: bool,
-    #[serde(default)]
-    pub zones: Option<Vec<MemoryZoneConfig>>,
 }
 
 impl MemoryConfig {
@@ -964,21 +840,6 @@ impl Default for MemoryConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct KernelConfig {
-    pub path: PathBuf,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct InitramfsConfig {
-    pub path: PathBuf,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Eq, Deserialize, Serialize)]
-pub struct CmdlineConfig {
-    pub args: String,
-}
-
 impl CmdlineConfig {
     pub fn parse(cmdline: Option<&str>) -> Result<Self> {
         let args = cmdline
@@ -987,41 +848,6 @@ impl CmdlineConfig {
 
         Ok(CmdlineConfig { args })
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct DiskConfig {
-    pub path: Option<PathBuf>,
-    #[serde(default)]
-    pub readonly: bool,
-    #[serde(default)]
-    pub direct: bool,
-    #[serde(default)]
-    pub iommu: bool,
-    #[serde(default = "default_diskconfig_num_queues")]
-    pub num_queues: usize,
-    #[serde(default = "default_diskconfig_queue_size")]
-    pub queue_size: u16,
-    #[serde(default)]
-    pub vhost_user: bool,
-    pub vhost_socket: Option<String>,
-    #[serde(default)]
-    pub rate_limiter_config: Option<RateLimiterConfig>,
-    #[serde(default)]
-    pub id: Option<String>,
-    // For testing use only. Not exposed in API.
-    #[serde(default)]
-    pub disable_io_uring: bool,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
-fn default_diskconfig_num_queues() -> usize {
-    DEFAULT_NUM_QUEUES_VUBLK
-}
-
-fn default_diskconfig_queue_size() -> u16 {
-    DEFAULT_QUEUE_SIZE_VUBLK
 }
 
 impl Default for DiskConfig {
@@ -1206,12 +1032,6 @@ impl DiskConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub enum VhostMode {
-    Client,
-    Server,
-}
-
 impl Default for VhostMode {
     fn default() -> Self {
         VhostMode::Client
@@ -1233,65 +1053,6 @@ impl FromStr for VhostMode {
             _ => Err(ParseVhostModeError::InvalidValue(s.to_owned())),
         }
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct NetConfig {
-    #[serde(default = "default_netconfig_tap")]
-    pub tap: Option<String>,
-    #[serde(default = "default_netconfig_ip")]
-    pub ip: Ipv4Addr,
-    #[serde(default = "default_netconfig_mask")]
-    pub mask: Ipv4Addr,
-    #[serde(default = "default_netconfig_mac")]
-    pub mac: MacAddr,
-    #[serde(default)]
-    pub host_mac: Option<MacAddr>,
-    #[serde(default)]
-    pub mtu: Option<u16>,
-    #[serde(default)]
-    pub iommu: bool,
-    #[serde(default = "default_netconfig_num_queues")]
-    pub num_queues: usize,
-    #[serde(default = "default_netconfig_queue_size")]
-    pub queue_size: u16,
-    #[serde(default)]
-    pub vhost_user: bool,
-    pub vhost_socket: Option<String>,
-    #[serde(default)]
-    pub vhost_mode: VhostMode,
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub fds: Option<Vec<i32>>,
-    #[serde(default)]
-    pub rate_limiter_config: Option<RateLimiterConfig>,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
-fn default_netconfig_tap() -> Option<String> {
-    None
-}
-
-fn default_netconfig_ip() -> Ipv4Addr {
-    Ipv4Addr::new(192, 168, 249, 1)
-}
-
-fn default_netconfig_mask() -> Ipv4Addr {
-    Ipv4Addr::new(255, 255, 255, 0)
-}
-
-fn default_netconfig_mac() -> MacAddr {
-    MacAddr::local_random()
-}
-
-fn default_netconfig_num_queues() -> usize {
-    DEFAULT_NUM_QUEUES_VUNET
-}
-
-fn default_netconfig_queue_size() -> u16 {
-    DEFAULT_QUEUE_SIZE_VUNET
 }
 
 impl Default for NetConfig {
@@ -1518,13 +1279,6 @@ impl NetConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct RngConfig {
-    pub src: PathBuf,
-    #[serde(default)]
-    pub iommu: bool,
-}
-
 impl RngConfig {
     pub fn parse(rng: &str) -> Result<Self> {
         let mut parser = OptionParser::new();
@@ -1553,17 +1307,6 @@ impl Default for RngConfig {
             iommu: false,
         }
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct BalloonConfig {
-    pub size: u64,
-    /// Option to deflate the balloon in case the guest is out of memory.
-    #[serde(default)]
-    pub deflate_on_oom: bool,
-    /// Option to enable free page reporting from the guest.
-    #[serde(default)]
-    pub free_page_reporting: bool,
 }
 
 impl BalloonConfig {
@@ -1602,28 +1345,6 @@ impl BalloonConfig {
             free_page_reporting,
         })
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct FsConfig {
-    pub tag: String,
-    pub socket: PathBuf,
-    #[serde(default = "default_fsconfig_num_queues")]
-    pub num_queues: usize,
-    #[serde(default = "default_fsconfig_queue_size")]
-    pub queue_size: u16,
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
-fn default_fsconfig_num_queues() -> usize {
-    1
-}
-
-fn default_fsconfig_queue_size() -> u16 {
-    1024
 }
 
 impl Default for FsConfig {
@@ -1707,21 +1428,6 @@ impl FsConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct PmemConfig {
-    pub file: PathBuf,
-    #[serde(default)]
-    pub size: Option<u64>,
-    #[serde(default)]
-    pub iommu: bool,
-    #[serde(default)]
-    pub discard_writes: bool,
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
 impl PmemConfig {
     pub const SYNTAX: &'static str = "Persistent memory parameters \
     \"file=<backing_file_path>,size=<persistent_memory_size>,iommu=on|off,\
@@ -1785,28 +1491,6 @@ impl PmemConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub enum ConsoleOutputMode {
-    Off,
-    Pty,
-    Tty,
-    File,
-    Null,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct ConsoleConfig {
-    #[serde(default = "default_consoleconfig_file")]
-    pub file: Option<PathBuf>,
-    pub mode: ConsoleOutputMode,
-    #[serde(default)]
-    pub iommu: bool,
-}
-
-fn default_consoleconfig_file() -> Option<PathBuf> {
-    None
-}
-
 impl ConsoleConfig {
     pub fn parse(console: &str) -> Result<Self> {
         let mut parser = OptionParser::new();
@@ -1864,17 +1548,6 @@ impl ConsoleConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct DeviceConfig {
-    pub path: PathBuf,
-    #[serde(default)]
-    pub iommu: bool,
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
 impl DeviceConfig {
     pub const SYNTAX: &'static str =
         "Direct device assignment parameters \"path=<device_path>,iommu=on|off,id=<device_id>,pci_segment=<segment_id>\"";
@@ -1923,15 +1596,6 @@ impl DeviceConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct UserDeviceConfig {
-    pub socket: PathBuf,
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
 impl UserDeviceConfig {
     pub const SYNTAX: &'static str =
         "Userspace device socket=<socket_path>,id=<device_id>,pci_segment=<segment_id>\"";
@@ -1974,23 +1638,6 @@ impl UserDeviceConfig {
 
         Ok(())
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct VdpaConfig {
-    pub path: PathBuf,
-    #[serde(default = "default_vdpaconfig_num_queues")]
-    pub num_queues: usize,
-    #[serde(default)]
-    pub iommu: bool,
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
-fn default_vdpaconfig_num_queues() -> usize {
-    1
 }
 
 impl VdpaConfig {
@@ -2052,18 +1699,6 @@ impl VdpaConfig {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct VsockConfig {
-    pub cid: u64,
-    pub socket: PathBuf,
-    #[serde(default)]
-    pub iommu: bool,
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub pci_segment: u16,
-}
-
 impl VsockConfig {
     pub const SYNTAX: &'static str = "Virtio VSOCK parameters \
         \"cid=<context_id>,socket=<socket_path>,iommu=on|off,id=<device_id>,pci_segment=<segment_id>\"";
@@ -2123,16 +1758,6 @@ impl VsockConfig {
 }
 
 #[cfg(target_arch = "x86_64")]
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct SgxEpcConfig {
-    pub id: String,
-    #[serde(default)]
-    pub size: u64,
-    #[serde(default)]
-    pub prefault: bool,
-}
-
-#[cfg(target_arch = "x86_64")]
 impl SgxEpcConfig {
     pub const SYNTAX: &'static str = "SGX EPC parameters \
         \"id=<epc_section_identifier>,size=<epc_section_size>,prefault=on|off\"";
@@ -2155,29 +1780,6 @@ impl SgxEpcConfig {
 
         Ok(SgxEpcConfig { id, size, prefault })
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct NumaDistance {
-    #[serde(default)]
-    pub destination: u32,
-    #[serde(default)]
-    pub distance: u8,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub struct NumaConfig {
-    #[serde(default)]
-    pub guest_numa_id: u32,
-    #[serde(default)]
-    pub cpus: Option<Vec<u8>>,
-    #[serde(default)]
-    pub distances: Option<Vec<NumaDistance>>,
-    #[serde(default)]
-    pub memory_zones: Option<Vec<String>>,
-    #[cfg(target_arch = "x86_64")]
-    #[serde(default)]
-    pub sgx_epc_sections: Option<Vec<String>>,
 }
 
 impl NumaConfig {
@@ -2266,58 +1868,6 @@ impl RestoreConfig {
             prefault,
         })
     }
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
-pub struct PayloadConfig {
-    #[serde(default)]
-    pub firmware: Option<PathBuf>,
-    #[serde(default)]
-    pub kernel: Option<PathBuf>,
-    #[serde(default)]
-    pub cmdline: Option<String>,
-    #[serde(default)]
-    pub initramfs: Option<PathBuf>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-pub struct VmConfig {
-    #[serde(default)]
-    pub cpus: CpusConfig,
-    #[serde(default)]
-    pub memory: MemoryConfig,
-    pub kernel: Option<KernelConfig>,
-    #[serde(default)]
-    pub initramfs: Option<InitramfsConfig>,
-    #[serde(default)]
-    pub cmdline: CmdlineConfig,
-    #[serde(default)]
-    pub payload: Option<PayloadConfig>,
-    pub disks: Option<Vec<DiskConfig>>,
-    pub net: Option<Vec<NetConfig>>,
-    #[serde(default)]
-    pub rng: RngConfig,
-    pub balloon: Option<BalloonConfig>,
-    pub fs: Option<Vec<FsConfig>>,
-    pub pmem: Option<Vec<PmemConfig>>,
-    #[serde(default = "ConsoleConfig::default_serial")]
-    pub serial: ConsoleConfig,
-    #[serde(default = "ConsoleConfig::default_console")]
-    pub console: ConsoleConfig,
-    pub devices: Option<Vec<DeviceConfig>>,
-    pub user_devices: Option<Vec<UserDeviceConfig>>,
-    pub vdpa: Option<Vec<VdpaConfig>>,
-    pub vsock: Option<VsockConfig>,
-    #[serde(default)]
-    pub iommu: bool,
-    #[cfg(target_arch = "x86_64")]
-    pub sgx_epc: Option<Vec<SgxEpcConfig>>,
-    pub numa: Option<Vec<NumaConfig>>,
-    #[serde(default)]
-    pub watchdog: bool,
-    #[cfg(feature = "guest_debug")]
-    pub gdb: bool,
-    pub platform: Option<PlatformConfig>,
 }
 
 impl VmConfig {
@@ -2747,6 +2297,7 @@ impl VmConfig {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use net_util::MacAddr;
 
     #[test]
     fn test_cpu_parsing() -> Result<()> {
